Fix gcc-11 warnings: signed/unsigned integer comparison, unused variables (#1463)
* Fix gcc-11 warnings: signed/unsigned integer comparision, unused variables


* declare loop variable i as unsigned int


* uncomment-out failable_free


* add commented-out assignment of failable_free as TODO


---------

+++ src/lib/OpenEXR/ImfIDManifestAttribute.cpp
@@ -40,7 +40,7 @@ IDManifestAttribute::readValueFrom (
     OPENEXR_IMF_INTERNAL_NAMESPACE::IStream& is, int size, int version)
 {
 
-    if (size < sizeof (uint64_t))
+    if (size < static_cast<int>(sizeof (uint64_t)))
     {
         throw IEX_NAMESPACE::InputExc (
             "Invalid size field reading idmanifest attribute");
